Move all TC output logic to TAP::Formatter::Session::TeamCity
This is how other alternative TAP formatters work, so this is the right way to
do formatting. This also simplifies some of the corner case handling a bit,
because we can respond to the close_test() event in the session and recover
from catastrophic death there, rather than at the start of the next test.