Get rid of all printing of raw output directly to STDOUT
In basically every case we were _also_ turning this into some sort of TC
message. This meant that this all ended up being doubled in the build log, and
it made for a big mess.

Now all test output is turned into some sort of TC message. Sometimes we
ignore output (if it's just whitespace or an empty comment line), and we
sometimes do a little cleanup on it too.