A horrible, ugly hack to solve TC slowness
With our internal test runners, the test grouping can change each run.
This causes TC to think that each test in a new group is a different
test. This hack is the simplest way to work around that. This is just a
temporary measure until the `Test2` formatter is ready, which should
provide a more robust solution.

THIS SHOULD NOT BE RELEASED TO CPAN.